TST: Running and fixing geo tests
